Bump version numbers so I won't time-travel backwards.
The internal (and since opensourcing) deprecated version had a release
between the creation of the opensource repository and its publication.
All the code is present, but I did not bump the version numbers.

This change will ensure I won't re-use a version number and thus safe
me the head-ache of going backwards in time/version-numbers.