Prevent deadlock when BufferedIterators are nested on the stack.
This makes them safe to be called from any ForkJoinPool, which is rather nice.

The original problem was that the BufferedIterator would block
(waiting for a wakeup), while the original thread already consumed all workers
in the threadpool.
As a consequence, the threadpool was exhausted and the job that would fulfill
the wakeup condition never started.
By calling the Wakeup blocking operation using ForkJoinPool.managedBlock(),
the pool will create new threads if required.