Replace gzipped flag with canAddSingleRecord.
Gzipped in v2.x files is an encoding property, but doesn't prohibit adding
records anymore.  So it makes no sense to base decisions off of that.
Instead, have files inform us they can are optimized (fast access) and if
they are suitable for adding single records.

I wouldn't be surprised if the v2.x format will have some more changes,
among which:
- index groups separately in List files.
- allow multiple time series in Tables files (allows appending bulk
  records after existing file).

If that happens, a canAddBulkRecords() bit will have to be created.