Remove iterator/stream wrappers from BufferedIterator.
The temptation is too large to just throw them in along a stream/iteration,
to split a computation over multiple threads (or deal with IO). But the
downside is that this will lead to deadlock scenarios, where all threads are
used for BufferedIterators, which are all waiting for other BufferedIterators
to do their work. And of course the latter can't, because they're waiting for
threads...